Externalize fixtures
