+++ tests/Model/BaseDocumentManagerTest.php
@@ -13,11 +13,7 @@ namespace Sonata\CoreBundle\Tests\Entity;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use PHPUnit\Framework\TestCase;
-use Sonata\CoreBundle\Model\BaseDocumentManager;
-
-class DocumentManager extends BaseDocumentManager
-{
-}
+use Sonata\CoreBundle\Tests\Fixtures\Model\DocumentManager;
 
 class BaseDocumentManagerTest extends TestCase
 {
