+++ tests/Model/BaseEntityManagerTest.php
@@ -14,11 +14,7 @@ namespace Sonata\CoreBundle\Tests\Entity;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\Common\Persistence\ObjectManager;
 use PHPUnit\Framework\TestCase;
-use Sonata\CoreBundle\Model\BaseEntityManager;
-
-class EntityManager extends BaseEntityManager
-{
-}
+use Sonata\CoreBundle\Tests\Fixtures\Entity\EntityManager;
 
 class BaseEntityManagerTest extends TestCase
 {
