+++ tests/Model/BaseManagerTest.php
@@ -12,29 +12,8 @@
 namespace Sonata\CoreBundle\Tests\Model;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
-use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\TestCase;
-use Sonata\CoreBundle\Model\BaseManager;
-
-class ManagerTest extends BaseManager
-{
-    /**
-     * Get the DB driver connection.
-     *
-     * @return Connection
-     */
-    public function getConnection()
-    {
-    }
-
-    /**
-     * @param $object
-     */
-    public function publicCheckObject($object)
-    {
-        return $this->checkObject($object);
-    }
-}
+use Sonata\CoreBundle\Tests\Fixtures\Model\Manager;
 
 /**
  * @author Hugo Briand <briand@ekino.com>
@@ -49,7 +28,7 @@ class BaseManagerTest extends TestCase
         $this->expectException(\InvalidArgumentException::class);
         $this->expectExceptionMessage('Object must be instance of class, DateTime given');
 
-        $manager = new ManagerTest('class', $this->createMock(ManagerRegistry::class));
+        $manager = new Manager('class', $this->createMock(ManagerRegistry::class));
 
         $manager->publicCheckObject(new \DateTime());
     }
